Move the configuration API under the api directory
This is probably more logically correct, and enables a simpler module
structure.
