+++ porch/hack/create-deployment-blueprint.sh
@@ -132,7 +132,7 @@ function main() {
   cp "${PORCH_DIR}/controllers/remoterootsync/config/rbac/role.yaml" \
      "${DESTINATION}/0-remoterootsync-role.yaml"
   # Repository CRD
-  cp "./controllers/pkg/apis/porch/v1alpha1/config.porch.kpt.dev_repositories.yaml" \
+  cp "./api/porchconfig/v1alpha1/config.porch.kpt.dev_repositories.yaml" \
      "${DESTINATION}/0-repositories.yaml"
 
   # Porch Deployment Config
